--- src/app/components/song-search/song-search.component.html
@@ -1,1 +1,0 @@
-<p>song-search works!</p>
