--- src/app/components/utils/search-bar/search-bar.component.html
@@ -1,1 +1,0 @@
-<p>search-bar works!</p>
